Set default issue sort
Issue sorting is required by further Jira queries but the
`userconfig_issuesort` is set only when configuring a project based on
`vira_projects.json`.  If projects are not defined, missing, or
whatever, then `userconfig_issuesort` is never set resulting in Jira
query errors.

This commit sets a default sane value to require users as little
configuration upfront as possible.